pyproject.toml
[project]
name = "spec-driven-delivery-tool"
version = "0.1.0"
description = "Lightweight internal web tool for requirements, scenarios, tasks, and acceptance tracking."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1.0",
    "openai>=2.26,<3.0",
    "pydantic-settings>=2.6,<3.0",
    "sqlmodel>=0.0.22,<0.1.0",
    "uvicorn[standard]>=0.32,<1.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28,<1.0",
    "pytest>=8.3,<9.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
